+++ Player.h
@@ -1,6 +1,9 @@
 #include <iostream>
 #include <vector>
 #include "Game.h"
+#include "horrorGame.h"
+#include "Simulator.h"
+#include "storyGame.h"
 #include "Achievements.h"
 #include <memory>
 #pragma once
@@ -10,14 +13,14 @@ using namespace std;
 class Player {
 private:
     vector<shared_ptr<Game>> games;
-    int money;
+    double money;
     Achievements ach;
     unsigned int goodGames, badGames;
 public:
     Player();
     Player(unsigned int);
     ~Player();
-    int getMoney() const;
+    double getMoney() const;
     void setMoney(int);
     void setGoodGames();
     void setBadGames();
